Refactor and small fix

+++ PhotonRenderer/Camera.cpp
@@ -25,8 +25,7 @@ namespace Photon
 		glm::mat4 projection = glm::mat4(1.0f);
 
 		view = glm::lookAt(m_position, m_position + m_rotation, m_up);
-		projection = glm::perspective(glm::radians(FOVdeg), (float)(m_width / m_height), near_plane, far_plane);
-		
+		projection = glm::perspective(glm::radians(FOVdeg), ((float)m_width / (float)m_height), near_plane, far_plane);
 		m_cam_matrix = projection * view;
 		m_view_matrix = view;
 	}
@@ -37,7 +36,7 @@ namespace Photon
 		glm::mat4 projection = glm::mat4(1.0f);
 
 		view = glm::lookAt(m_position, m_position + m_rotation, m_up);
-		projection = glm::perspective(glm::radians(m_fov), (float)(m_width / m_height), m_near_plane, m_far_plane);
+		projection = glm::perspective(glm::radians(m_fov), (m_width / (float)m_height), m_near_plane, m_far_plane);
 		//projection = glm::ortho(0, m_width, m_height, 0, 0, 1000);
 		m_cam_matrix = projection * view;
 		m_view_matrix = view;
